--- a/stoplight.gemspec
+++ b/stoplight.gemspec
@@ -28,6 +28,8 @@ Gem::Specification.new do |gem|
 
   gem.required_ruby_version = '>= 1.9.3'
 
+  gem.add_dependency 'concurrent-ruby'
+
   {
     'benchmark-ips' => '2.3',
     'bugsnag' => '2.8',