@tfausak 2016-02-12
This looks good! I'm not sure why the build failed. Doesn't seem like a problem with this branch. 

I wonder if the thread-safe data store should be provided as an option, much like the Redis data store. That way I can avoid adding `concurrent-ruby` as a dependency, and only users that need thread safety have to pay for it. What do you think?