@AndreasWurm 2016-02-12
well, i think thats your decision, only point i would like to raise is to remember the fallback logic, when redis fails, the Light will fallback to use Default::DATA_STORE.