@tfausak 2016-02-15
That's a good point. Can you run [the benchmarks](https://github.com/orgsync/stoplight/blob/v1.3.0/bench/memory_bench.rb) against this branch?