@tfausak 2016-02-18
The benchmarks don't test the failure code path, but they run in about the same amount of time as on master.

```
# master
     creating lambda:  1647203.1 i/s
      calling lambda:  1381496.4 i/s - same-ish: difference falls within error
  creating stoplight:   544997.1 i/s - 3.02x slower
   calling stoplight:   170786.9 i/s - 9.64x slower
```

```
# this branch
     creating lambda:  1607600.8 i/s
      calling lambda:  1464719.6 i/s - same-ish: difference falls within error
  creating stoplight:   555682.0 i/s - 2.89x slower
   calling stoplight:   165307.7 i/s - 9.72x slower
```